JAVA-3640: move accumulator methods to Accumulators
