+++ driver-scala/src/main/scala/org/mongodb/scala/model/Aggregates.scala
@@ -42,124 +42,6 @@ object Aggregates {
   def addFields(fields: Field[_]*): Bson = JAggregates.addFields(fields.asJava)
 
   /**
-   * Creates an \$accumulator pipeline stage
-   *
-   * @param initFunction         a function used to initialize the state
-   * @param accumulateFunction   a function used to accumulate documents
-   * @param mergeFunction        a function used to merge two internal states, e.g. accumulated on different shards or
-   *                             threads. It returns the resulting state of the accumulator.
-   * @return the \$accumulator pipeline stage
-   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/accumulator/ \$accumulator]]
-   * @since 1.2
-   * @note Requires MongoDB 4.4 or greater
-   */
-  def accumulator(
-      initFunction: String,
-      accumulateFunction: String,
-      mergeFunction: String
-  ): Bson =
-    JAggregates.accumulator(
-      initFunction,
-      accumulateFunction,
-      mergeFunction
-    )
-
-  /**
-   * Creates an \$accumulator pipeline stage
-   *
-   * @param initFunction         a function used to initialize the state
-   * @param accumulateFunction   a function used to accumulate documents
-   * @param mergeFunction        a function used to merge two internal states, e.g. accumulated on different shards or
-   *                             threads. It returns the resulting state of the accumulator.
-   * @param finalizeFunction     a function used to finalize the state and return the result (may be null)
-   * @return the \$accumulator pipeline stage
-   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/accumulator/ \$accumulator]]
-   * @since 1.2
-   * @note Requires MongoDB 4.4 or greater
-   */
-  def accumulator(
-      initFunction: String,
-      accumulateFunction: String,
-      mergeFunction: String,
-      finalizeFunction: String
-  ): Bson =
-    JAggregates.accumulator(
-      initFunction,
-      accumulateFunction,
-      mergeFunction,
-      finalizeFunction
-    )
-
-  /**
-   * Creates an \$accumulator pipeline stage
-   *
-   * @param initFunction         a function used to initialize the state
-   * @param initArgs             init function’s arguments (may be null)
-   * @param accumulateFunction   a function used to accumulate documents
-   * @param accumulateArgs       additional accumulate function’s arguments (may be null). The first argument to the
-   *                             function is ‘state’.
-   * @param mergeFunction        a function used to merge two internal states, e.g. accumulated on different shards or
-   *                             threads. It returns the resulting state of the accumulator.
-   * @param finalizeFunction     a function used to finalize the state and return the result (may be null)
-   * @return the \$accumulator pipeline stage
-   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/accumulator/ \$accumulator]]
-   * @since 1.2
-   * @note Requires MongoDB 4.4 or greater
-   */
-  def accumulator(
-      initFunction: String,
-      initArgs: Seq[String],
-      accumulateFunction: String,
-      accumulateArgs: Seq[String],
-      mergeFunction: String,
-      finalizeFunction: String
-  ): Bson =
-    JAggregates.accumulator(
-      initFunction,
-      initArgs.asJava,
-      accumulateFunction,
-      accumulateArgs.asJava,
-      mergeFunction,
-      finalizeFunction
-    )
-
-  /**
-   * Creates an \$accumulator pipeline stage
-   *
-   * @param initFunction         a function used to initialize the state
-   * @param initArgs             init function’s arguments (may be null)
-   * @param accumulateFunction   a function used to accumulate documents
-   * @param accumulateArgs       additional accumulate function’s arguments (may be null). The first argument to the
-   *                             function is ‘state’.
-   * @param mergeFunction        a function used to merge two internal states, e.g. accumulated on different shards or
-   *                             threads. It returns the resulting state of the accumulator.
-   * @param finalizeFunction     a function used to finalize the state and return the result (may be null)
-   * @param lang                 a language specifier
-   * @return the \$accumulator pipeline stage
-   * @see [[http://docs.mongodb.org/manual/reference/operator/aggregation/accumulator/ \$accumulator]]
-   * @since 1.2
-   * @note Requires MongoDB 4.4 or greater
-   */
-  def accumulator(
-      initFunction: String,
-      initArgs: Seq[String],
-      accumulateFunction: String,
-      accumulateArgs: Seq[String],
-      mergeFunction: String,
-      finalizeFunction: String,
-      lang: String
-  ): Bson =
-    JAggregates.accumulator(
-      initFunction,
-      initArgs.asJava,
-      accumulateFunction,
-      accumulateArgs.asJava,
-      mergeFunction,
-      finalizeFunction,
-      lang
-    )
-
-  /**
    * Creates a \$bucket pipeline stage
    *
    * @param groupBy    the criteria to group By
